Fix warnings in examples

+++ examples/10_calling_functions.cpp
@@ -29,7 +29,7 @@ struct FunctionBArguments
 
 struct FunctionBReturn
 {
-    int functionBReturnA;
+    float functionBReturnA;
     std::string functionBReturnB;
     double functionBReturnC[3];
